requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.3
django-ckeditor==5.9.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-jsonfield==1.4.0
django-materialize==1.0.1a2
django-materializecss-form==1.1.17
django-stripe-payments==2.0.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.8
Pillow==7.0.0
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
stripe==2.42.0
urllib3==1.25.8
whitenoise==5.0.1
